requirements.txt
Flask==2.2.5
gTTS==2.3.2
keras==2.13.1
numpy==1.24.3
opencv_python==4.8.0.76
Pillow==10.0.0
Pillow==10.1.0
tensorflow==2.13.0
tensorflow_intel==2.13.0

